requirements.txt
# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Encryption
cryptography==41.0.8
python-dotenv==1.0.0

# Connection pooling
sqlalchemy[asyncio]==2.0.23

# Development
pytest==7.4.3
pytest-asyncio==0.21.1

# Authentication & Security
bcrypt==4.1.2
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6